eval.py bug fixed
Some modifications were made in the new version of eval.py, but removing the logger will cause an error in some environments.
MisconfigurationException : Cannot use `LearningRateMonitor` callback with `Trainer` that has no logger.
Another suggestion: modify the configuration file as follows to avoid warning log outputs:
optimize_SMPL:
      enable: true 
      is_refine: true
It seems that your method in the paper also needs to be set to true. I'm not sure why the demo configuration file has it set to false.